requirements.txt
Django==1.8
django-allauth==0.24.1
django-analytical==2.2.1
django-dajaxice==0.7
django-guardian==1.3.1
django-widget-tweaks==1.4.1
gunicorn==19.4.1
oauthlib==1.0.3
Pillow==3.0.0
psycopg2==2.6.1
PyJWT==1.4.0
python-openid==2.2.5
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
